# apps/api/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SurauResponse": {
				"properties": {
					"brief_direction": {
						"type": "string"
					},
					"mall": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"is_approved_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_approved": {
						"type": "boolean"
					},
					"unique_name": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"brief_direction",
					"mall",
					"district",
					"state",
					"is_approved_at",
					"created_at",
					"is_approved",
					"unique_name",
					"name"
				],
				"type": "object"
			},
			"Surau": {
				"properties": {
					"is_solat_jumaat": {
						"type": "boolean"
					},
					"user_id": {
						"type": "string"
					},
					"is_qiblat_certified": {
						"type": "boolean"
					},
					"brief_direction": {
						"type": "string"
					},
					"mall": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"is_approved_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_approved": {
						"type": "boolean"
					},
					"unique_name": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"is_solat_jumaat",
					"user_id",
					"is_qiblat_certified",
					"brief_direction",
					"mall",
					"district",
					"state",
					"is_approved_at",
					"created_at",
					"is_approved",
					"unique_name",
					"name"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "carisurau-api",
		"version": "0.0.1",
		"description": "Cari surau API",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Farhan Helmy",
			"email": "farhanhlmy@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/surau": {
			"get": {
				"operationId": "GetSurau",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SurauResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "PostSurau",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SurauResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Surau"
							}
						}
					}
				}
			}
		},
		"/surau/{id}": {
			"get": {
				"operationId": "GetOneSurau",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SurauResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchSurau",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SurauResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteSurau",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SurauResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rating/{id}": {
			"delete": {
				"operationId": "DeleteRating",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}