# apps/api/src/types/surau.ts
export type Surau = {
    name: string
    unique_name: string
    is_approved: boolean
    created_at: Date
    is_approved_at: Date
    state: string
    district: string
    mall: string
    brief_direction: string
    is_qiblat_certified: boolean
    user_id: string
    is_solat_jumaat: boolean
} | undefined